+++ src/labrinth/calls.rs
@@ -1,61 +1,47 @@
-//! his file contains abstractions of some of the calls supported by the Labrinth API
+//! This file contains abstractions of some of the calls supported by the Labrinth API
 
 use super::structs::*;
+use crate::ferium_error::{FError, FResult};
 use bytes::Bytes;
-use reqwest::StatusCode;
 use reqwest::{Client, Response};
 
 /// Return the contents of `version`'s JAR file as bytes
-pub async fn download_version(client: &Client, version: &Version) -> Bytes {
-    request(client, version.files[0].url.clone())
-        .await
+pub async fn download_version(client: &Client, version: Version) -> FResult<Bytes> {
+    Ok(request(client, version.files[0].url.clone())
+        .await?
         .bytes()
-        .await
-        .unwrap()
-}
-
-/// Checks if a mod exists. If it does, then the mod is returned, else None is returned
-pub async fn does_exist(client: &Client, mod_id: &ID) -> Option<Mod> {
-    let response = request_rel(client, format!("/mod/{}", mod_id)).await;
-    match response.status() {
-        StatusCode::OK => response.json().await.unwrap(),
-        _ => Option::None,
-    }
+        .await?)
 }
 
 /// Returns the versions of `mod_id`'s mod sorted in chronologically descending order
-pub async fn get_versions(client: &Client, mod_id: &str) -> Vec<Version> {
-    request_rel(client, format!("/mod/{}/version", mod_id))
-        .await
+pub async fn get_versions(client: &Client, mod_id: &str) -> FResult<Vec<Version>> {
+    Ok(request_rel(client, format!("/mod/{}/version", mod_id))
+        .await?
         .json()
-        .await
-        .unwrap()
+        .await?)
 }
 
 /// Get a mod using the `mod_slug`, which can also be the mod ID
-pub async fn get_mod(client: &Client, mod_slug: &str) -> Mod {
-    request_rel(client, format!("/mod/{}", mod_slug))
-        .await
+pub async fn get_mod(client: &Client, mod_slug: &str) -> FResult<Mod> {
+    Ok(request_rel(client, format!("/mod/{}", mod_slug))
+        .await?
         .json()
-        .await
-        .unwrap()
+        .await?)
 }
 
 /// Send a request to `url` with `client` and return response. Labrinth's base URL will be prepended to `url`
-pub async fn request_rel(client: &Client, url: String) -> Response {
-    request(
-        client,
-        format!("https://api.modrinth.com/api/v1{}", url).into(),
-    )
-    .await
+async fn request_rel(client: &Client, url: String) -> FResult<Response> {
+    Ok(request(client, format!("https://api.modrinth.com/api/v1{}", url)).await?)
 }
 
 /// Send a request to `url` with `client` and return response
-pub async fn request(client: &Client, url: String) -> Response {
-    let response = client.get(url).send().await.unwrap();
+async fn request(client: &Client, url: String) -> FResult<Response> {
+    let response = client.get(url).send().await?;
     if response.status().is_success() {
-        response
+        Ok(response)
     } else {
-        panic!("HTTP request failed with error code {}", response.status());
+        Err(FError::HTTPError {
+            message: format!("HTTP request failed with error code {}", response.status()),
+        })
     }
 }