+++ src/util/mod.rs
@@ -1,3 +1,5 @@
 pub mod cli;
+pub mod ferium_error;
 pub mod json;
+pub mod launchermeta;
 pub mod wrappers;